artifacts-service/pyproject.toml
[project]
name = "artifacts-service"
version = "0.1.0"
description = "Artifacts service for core AI"
authors = [
    { name = "Bbar0n234", email = "vilnus1337@mail.ru" }
]
license = {text = "Apache 2.0"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "uvicorn[standard]>=0.35.0",
    "markdown>=3.5",
    "python-markdown-math>=0.8",
    "pdfkit>=1.0.0",
    "aiofiles>=23.2.1",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "python-jose[cryptography]>=3.3.0",
    # Web3 authentication dependencies
    "eth-account>=0.10.0",
    "web3>=6.0.0",
    "pyjwt>=2.8.0",
    # HTTP client for proxy requests
    "httpx>=0.25.0",
    # Form data support for FastAPI
    "python-multipart>=0.0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.17.1",
    "ruff>=0.12.7",
]
test = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
